Fixed residual bugs in the autonomous framework.

+++ comp2010/auton.h
@@ -12,9 +12,9 @@
 #define TURN_ERROR     ((int16_t)30 * ENC_PER_DEG)
 #define TURN_THRESHOLD ((int16_t)5  * ENC_PER_DEG)
 
-#define SLOW_US 18500
+#define SLOW_US ((uint32_t)18500)
 
-#define MS_TO_LOOPS(_x_)   (((_x_) * (uint32_t)1000 / SLOW_US))
+#define MS_TO_LOOPS(_x_)   (((uint32_t)(_x_) * (uint32_t)1000 / SLOW_US))
 #define IN10_TO_TICKS(_x_) ((uint32_t)(_x_) * ENC_PER_IN / 10)
 #define DEG_TO_TICKS(_x_)  ((uint32_t)(_x_) * ENC_PER_DEG)
